// src/generated/models/Crf82_leaverequestsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Crf82_leaverequestscrf82_leavetype = {
  0: 'CP',
  1: 'RTT',
  2: 'Remotework'
} as const;
export type Crf82_leaverequestscrf82_leavetype = keyof typeof Crf82_leaverequestscrf82_leavetype;
export const Crf82_leaverequestscrf82_requeststatus = {
  0: 'Pending',
  1: 'Accepted',
  2: 'Refused'
} as const;
export type Crf82_leaverequestscrf82_requeststatus = keyof typeof Crf82_leaverequestscrf82_requeststatus;
export const Crf82_leaverequestsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Crf82_leaverequestsstatecode = keyof typeof Crf82_leaverequestsstatecode;
export const Crf82_leaverequestsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Crf82_leaverequestsstatuscode = keyof typeof Crf82_leaverequestsstatuscode;

export interface Crf82_leaverequestsBase {
  crf82_comments?: string;
  crf82_creationtimestamp?: string;
  crf82_decisionmaker?: string;
  crf82_decisiontimestamp?: string;
  crf82_enddate?: string;
  crf82_leaverequestid: string;
  crf82_leavetype?: Crf82_leaverequestscrf82_leavetype;
  crf82_legacyidentifier?: number;
  crf82_requestername?: string;
  crf82_requeststatus?: Crf82_leaverequestscrf82_requeststatus;
  crf82_requesttitle: string;
  crf82_startdate?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Crf82_leaverequestsstatecode;
  statuscode?: Crf82_leaverequestsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Crf82_leaverequests extends Crf82_leaverequestsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  crf82_leavetypename?: string;
  crf82_requeststatusname?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}